Fix cyclical import issue
